+++ frontend/src/pages/PRCommentPage.module.css
@@ -93,7 +93,7 @@
   border-radius: 21px;
   border: 1px solid lightgray;
   box-shadow: 5px 5px 5px #d3d3d37a;
-  width: 60%;
+  width: 100%;
   margin: auto;
 }
 .commentList .commentUser {
@@ -147,19 +147,12 @@ pre {
 pre > div {
   white-space: pre-wrap;
 }
-.prSidebar {
-  height: 35rem;
-  overflow: auto;
-  border: 1px solid #dadada;
-  padding: 1rem;
-  border-radius: 15px 0 0 15px;
-  margin-top: 2.5rem;
-}
 .lineBody {
   color: black;
   font-size: 16px;
 }
 .codeLineWrapper .commentBody {
+  width: 60%;
   padding: 1rem;
   margin-bottom: 1rem;
 }
